Appearance.kt

+++ app/src/main/java/com/mike/studentportal/Settings.kt
@@ -122,7 +122,7 @@ fun SettingsScreen(navController: NavController, context: Context) {
                     MyDatabase.getScreenTime(screenID) { existingScreenTime ->
                         val totalScreenTime = if (existingScreenTime != null) {
                             Log.d("Screen Time", "Retrieved Screen time: $existingScreenTime")
-                            existingScreenTime.time.toLong() + timeSpent
+                            existingScreenTime.time + timeSpent
                         } else {
                             timeSpent
                         }
